Fix Bow Wow pole start sequences: Photo, Take for Walk
To trigger the sequences for Bow Wow, you need to touch the pole. The ranges were not very good since to get the photo, only touching the right side of the pole worked. To take him for a walk, only touching the front of the pole worked. Now touching any part of the pole will trigger both sequences.