Fix Marin dungeon sequence (a.k.a. longstanding Marin beach issue).
This is at long last hopefully the fix to the Marin beach spawning issue that has plagued this port since the v1.0.0 release. I created a workaround awhile back that seemed to work for the most part, but the issue recently returned and revealed itself to be the dungeon sequence firing when it shouldn't have been.

In earlier versions, Marin would sometimes just straight up disappear after the beach sequence. I created a workaround that respawned her object. Marin is now an "always animated" object, so she can still be rendered and updated outside the screen boundaries. The bug has returned and revealed itself to be the dungeon sequence: aka, she spawns next to the dungeon waiting for Link to come out.

This commit hopefully finally fixes the issue once and for all. It checks if the follower key is set before trying to fire any logic instead of just relying on whether or not her object is currently on the map (because it's always on the map). This also seems to fix another issue where saving outside a dungeon, saving, quitting, and reloading the save would sometimes not respawn Marin. So hopefully all Marin related issues are finally fixed for good.