Greatly reduce underwater knockback effects.
It appears that water amplifies the knockback effects. Everything from taking damage, to hitting with a sword, to being repelled would blow the player to the edge of the screen. A few versions back, knockback effects were increased to apply greater velocity but decay quicker. Water effects do not decay nearly as quickly.