Fix Rapids Ride waterfalls, raft clipping, & rooster exploit.
This is an area of the map I mostly neglected during my time working on this. Fixes being able to clip off the raft with the right maneuvers when pushing against the ladder/water diagonally then switching to a downward position real fast. It's also now possible to sneak into Rapids Ride (the rafting area) with the flying rooster and a secret method. But this area was never designed for Link to be in there without a raft. There were parts where "deep water" was not applied, waterfalls were not equipped with "Rail Jump" objects and had no collision at the top, so he would just walk on the waterfall and get stuck. Trying to enter the raft while flying with the rooster would bug out Link, so prevent that from happening as well.