Rework grabbing/pulling animations and mechanics.
This reworks how grabbing and pulling works and also updates the animations so that they match the original game. It exploits the geometry by using sub-pixels to extend the grab rectangle into an object so that grabs can look more accurate. When Link was facing down and grabbing a wall, there was a 1 pixel gap between him and the wall, so the grab box was extended with sub-pixels to reach just a little bit further. For example, collision is 16x16 but the grabbable box is 14x14. Link's grab box at 2x2 only reaches into the collision by 1 pixel, the other is inside his body. So the grab box is extended by 0.04 units on each side, which is enough to trigger a grab but not enough to trigger it early. It's really hard to explain thoroughly, but it works well so that's what really matters.