Shopkeeper name buyback removes a death.
This was suggested to me a long time ago but I decided against it. But I thought about it and yeah, maybe it would be nice to have a way to remove the mandatory death he gives you if you want to get all the photos. It still requires "Nothing is Missable" since it changes base game behavior, but at least it's an option now.