Fix several issues with enemy Karakoro (Color Dungeon).
This is the ball like creature you throw into holes and match their colors. In the original game, it was possible to walk on top of them when they were in a hole so restore that behavior. Also fix resetting their states in Classic Camera when a field transition happens.